Update Mk5Quarry.cs
I placed comments in the code for the sections that I changed and Why.  Can take the comments out if end up using the changes they where just put in place to show why each change was made. To summarize, made changes in the section for mining garbage, so it will unload every garbage block if config has mining of garbage set.  Fix to the logic of the statement that checks if an ore should be skipped,  Before had to have both AllOre and the ore type in the config to get it to actually mine the block.  Added code so that when Ablated Resin if found, it treats it the same as Hardened Resin.  Last change was to change from true to false the return on Resin handling as the function needs to return false for the quarry to try and offload the resin.